requirements.txt
smitheory-ai
requests
python-dotenv
flask
gunicorn